crt: modify mtvec
It turned out that mtvec is not hard-wired to 0xc0. Higher bits can be
modified to keep trap_entry location is in sram/eeprom/spifi. So no need
to change boot mode jumpers just to map trap_handler to address 0xc0.

This commit writes trap_entry address to mtvec. This will not change
lower 0xc0 bits, but it will change higher bits. Note that lower part
is guaranteed to be correct since trap_entry is placed in properly
positioned .trap section.

For instance, see openocd commands:

: > reg mtvec
: mtvec (/32): 0x800000c0
: > reg mtvec 0x020000c0
: mtvec (/32): 0x020000c0
: > reg mtvec
: mtvec (/32): 0x020000c0

Signed-off-by: Sergey Matyukevich <geomatsi@gmail.com>